refactor: internal_handlers_folder
